requirements.txt
# Core
Django
djangorestframework
django-cors-headers
python-dotenv

# Background jobs
celery
redis

# CV / video
opencv-python-headless==4.10.0.84
python-dotenv~=1.0.0
numpy>=2.0,<2.3
moviepy
ffmpeg-python
numpy

# Deep learning (CPU default wheel).
# If you need CUDA, see note below to install from the PyTorch index.
torch
torchvision

# Jupyter kernel for this venv
ipykernel

# Inference toolkit with extras
# (transformers, SAM, CLIP, gaze, GroundingDINO)
inference[transformers,sam,clip,gaze,grounding-dino]

# Football Analytics
ultralytics
matplotlib
tqdm
timm
transformers
sentencepiece
gdown
supervision
more_itertools
scikit-learn
umap-learn